bcs and extent now part of the Grid not Mesh class
