+++ src/framework/domain/mesh.h
@@ -16,7 +16,6 @@
 #include "enums.h"
 #include "global.h"
 
-#include "arch/directions.h"
 #include "utils/comparators.h"
 #include "utils/error.h"
 #include "utils/numeric.h"
@@ -35,37 +34,28 @@ namespace ntt {
   template <class M>
     requires metric::traits::HasD<M> && metric::traits::HasConvert_i<M>
   struct Mesh : public Grid<M::Dim> {
-    static constexpr bool      is_mesh { true };
     static constexpr Dimension D { M::Dim };
+    using base_t = Grid<D>;
+    using base_t::extent;
+    using base_t::m_extent;
+    using base_t::m_flds_bc;
+    using base_t::m_prtl_bc;
 
     M metric;
 
     Mesh(const std::vector<ncells_t>&         res,
          const boundaries_t<real_t>&          ext,
          const std::map<std::string, real_t>& metric_params)
-      : Grid<D> { res }
-      , metric { res, ext, metric_params }
-      , m_extent { ext } {}
+      : Grid<D> { res, ext }
+      , metric { res, ext, metric_params } {}
 
     Mesh(const std::vector<ncells_t>&         res,
          const boundaries_t<real_t>&          ext,
          const std::map<std::string, real_t>& metric_params,
          const boundaries_t<FldsBC>&          flds_bc,
          const boundaries_t<PrtlBC>&          prtl_bc)
-      : Grid<D> { res }
-      , metric { res, ext, metric_params }
-      , m_extent { ext } {
-      for (auto d { 0 }; d < D; ++d) {
-        dir::direction_t<D> dir_plus;
-        dir_plus[d] = +1;
-        dir::direction_t<D> dir_minus;
-        dir_minus[d] = -1;
-        set_flds_bc(dir_plus, flds_bc[d].second);
-        set_flds_bc(dir_minus, flds_bc[d].first);
-        set_prtl_bc(dir_plus, prtl_bc[d].second);
-        set_prtl_bc(dir_minus, prtl_bc[d].first);
-      }
-    }
+      : Grid<D> { res, ext, flds_bc, prtl_bc }
+      , metric { res, ext, metric_params } {}
 
     ~Mesh() = default;
 
@@ -195,106 +185,6 @@ namespace ntt {
       }
       return range;
     }
-
-    /* getters -------------------------------------------------------------- */
-    [[nodiscard]]
-    auto extent(in i) const -> std::pair<real_t, real_t> {
-      switch (i) {
-        case in::x1:
-          return (m_extent.size() > 0) ? m_extent[0]
-                                       : std::pair<real_t, real_t> { ZERO, ZERO };
-        case in::x2:
-          return (m_extent.size() > 1) ? m_extent[1]
-                                       : std::pair<real_t, real_t> { ZERO, ZERO };
-        case in::x3:
-          return (m_extent.size() > 2) ? m_extent[2]
-                                       : std::pair<real_t, real_t> { ZERO, ZERO };
-        default:
-          raise::Error("invalid dimension", HERE);
-          throw;
-      }
-    }
-
-    [[nodiscard]]
-    auto extent() const -> boundaries_t<real_t> {
-      return m_extent;
-    }
-
-    [[nodiscard]]
-    auto flds_bc() const -> boundaries_t<FldsBC> {
-      if constexpr (D == Dim::_1D) {
-        return {
-          { flds_bc_in({ -1 }), flds_bc_in({ 1 }) }
-        };
-      } else if constexpr (D == Dim::_2D) {
-        return {
-          { flds_bc_in({ -1, 0 }), flds_bc_in({ 1, 0 }) },
-          { flds_bc_in({ 0, -1 }), flds_bc_in({ 0, 1 }) }
-        };
-      } else if constexpr (D == Dim::_3D) {
-        return {
-          { flds_bc_in({ -1, 0, 0 }), flds_bc_in({ 1, 0, 0 }) },
-          { flds_bc_in({ 0, -1, 0 }), flds_bc_in({ 0, 1, 0 }) },
-          { flds_bc_in({ 0, 0, -1 }), flds_bc_in({ 0, 0, 1 }) }
-        };
-      } else {
-        raise::Error("invalid dimension", HERE);
-        throw;
-      }
-    }
-
-    [[nodiscard]]
-    auto prtl_bc() const -> boundaries_t<PrtlBC> {
-      if constexpr (D == Dim::_1D) {
-        return {
-          { prtl_bc_in({ -1 }), prtl_bc_in({ 1 }) }
-        };
-      } else if constexpr (D == Dim::_2D) {
-        return {
-          { prtl_bc_in({ -1, 0 }), prtl_bc_in({ 1, 0 }) },
-          { prtl_bc_in({ 0, -1 }), prtl_bc_in({ 0, 1 }) }
-        };
-      } else if constexpr (D == Dim::_3D) {
-        return {
-          { prtl_bc_in({ -1, 0, 0 }), prtl_bc_in({ 1, 0, 0 }) },
-          { prtl_bc_in({ 0, -1, 0 }), prtl_bc_in({ 0, 1, 0 }) },
-          { prtl_bc_in({ 0, 0, -1 }), prtl_bc_in({ 0, 0, 1 }) }
-        };
-      } else {
-        raise::Error("invalid dimension", HERE);
-        throw;
-      }
-    }
-
-    [[nodiscard]]
-    auto flds_bc_in(const dir::direction_t<D>& direction) const -> FldsBC {
-      raise::ErrorIf(m_flds_bc.find(direction) == m_flds_bc.end(),
-                     "direction not found",
-                     HERE);
-      return m_flds_bc.at(direction);
-    }
-
-    [[nodiscard]]
-    auto prtl_bc_in(const dir::direction_t<D>& direction) const -> PrtlBC {
-      raise::ErrorIf(m_prtl_bc.find(direction) == m_prtl_bc.end(),
-                     "direction not found",
-                     HERE);
-      return m_prtl_bc.at(direction);
-    }
-
-    /* setters -------------------------------------------------------------- */
-    inline void set_flds_bc(const dir::direction_t<D>& direction, const FldsBC& bc) {
-      m_flds_bc.insert_or_assign(direction, bc);
-    }
-
-    inline void set_prtl_bc(const dir::direction_t<D>& direction, const PrtlBC& bc) {
-      m_prtl_bc.insert_or_assign(direction, bc);
-    }
-
-  private:
-    boundaries_t<real_t>  m_extent;
-    dir::map_t<D, FldsBC> m_flds_bc;
-    dir::map_t<D, PrtlBC> m_prtl_bc;
   };
 } // namespace ntt
